pull now playing logic out of YTSession
